rename Agilent E8267 driver
